+++ pkg/apiserver/apiserver_test.go
@@ -24,6 +24,8 @@ import (
 	"net/http/httptest"
 	"reflect"
 	"testing"
+
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/labels"
 )
 
 // TODO: This doesn't reduce typing enough to make it worth the less readable errors. Remove.
@@ -49,7 +51,7 @@ type SimpleRESTStorage struct {
 	updated Simple
 }
 
-func (storage *SimpleRESTStorage) List(LabelQuery) (interface{}, error) {
+func (storage *SimpleRESTStorage) List(labels.Query) (interface{}, error) {
 	result := SimpleList{
 		Items: storage.list,
 	}
